+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2023-08-23
+
+### Added
+
+- Added Longitude and Laditude
+
+### Changed
+
+- Updated OpenWeather API to use longitude and laditude instead of city name
+
 ## [2.1.2] - 2020-06-25
 
 ### Added